executor: move some test into writetest
