+++ executor/writetest/BUILD.bazel
@@ -1,0 +1,38 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "writetest_test",
+    srcs = [
+        "main_test.go",
+        "write_test.go",
+    ],
+    flaky = True,
+    race = "on",
+    shard_count = 50,
+    deps = [
+        "//config",
+        "//executor",
+        "//kv",
+        "//meta/autoid",
+        "//parser/model",
+        "//parser/mysql",
+        "//planner/core",
+        "//session",
+        "//sessionctx",
+        "//sessionctx/stmtctx",
+        "//sessionctx/variable",
+        "//sessiontxn",
+        "//store/mockstore",
+        "//table",
+        "//table/tables",
+        "//testkit",
+        "//types",
+        "//util",
+        "//util/mock",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@io_opencensus_go//stats/view",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
